Fixed: 실시간 투표 기능
유저 Id를 Body데이터로 받아 처리하도록 수정

+++ src/main/java/fish/common/community/vote/service/VoteService.java
@@ -17,15 +17,18 @@ public class VoteService {
     private final VoteRepository voteRepository;
 
     @Transactional
-    public List<Map<String, Object>> saveVote(Long postId, Long userId, VoteRequest request) {
-        Optional<VoteEntity> vote = voteRepository.findByPostIdAndUserId(postId, userId);
-        // 중복 투표의 경우 기존 투표 취소
-        if (vote.isPresent()) {
+    public List<Map<String, Object>> saveVote(Long postId, VoteRequest request) {
+        Optional<VoteEntity> vote = voteRepository.findByPostIdAndUserId(postId, request.getUserId());
+
+        // 1. 이미 동일한 투표를 했으면 취소 (삭제)
+        if (vote.isPresent() && vote.get().getVoteOption().equals(request.getVoteOption())) {
             voteRepository.delete(vote.get());
+            return voteRepository.findVoteCountsByPostId(postId);
         }
-        else {
-            voteRepository.save(VoteEntity.toEntity(postId, userId, request));
-        }
+
+        // 2. 기존에 다른 선택지를 투표한 경우 기존 투표 삭제 후 새 투표 저장
+        vote.ifPresent(voteRepository::delete);
+        voteRepository.save(VoteEntity.toEntity(postId, request.getUserId(), request));
 
         // 현재 집계 수 소켓 메세지로 보내주기
         return voteRepository.findVoteCountsByPostId(postId);
